+++ R/input_check_response.R
@@ -1,27 +1,31 @@
-#' Standardized response to an input check
+#' Standardized response to input check
 #'
 #' @description
-#' This function provides standardized responses to input checks, ensuring
+#' This function provides a standardized response to input checks, ensuring
 #' consistency.
 #'
-#' @param check \[`TRUE` | `character(1)`\]\cr
+#' @param check \[`TRUE` | `character(1)` | `list()`\]\cr
 #' Matches the return value of the `check*` functions from the `{checkmate}`
 #' package, i.e., either `TRUE` if the check was successful, or a `character`
 #' (the error message) else.
+#'
+#' Can also be a `list` of multiple such values for alternative criteria, where
+#' at least one must be `TRUE` for a successful check.
 #'
 #' @param var_name \[`NULL` | `character(1)`\]\cr
 #' Optionally specifies the name of the input being checked. This name will be
 #' used for the default value of the `prefix` argument.
 #'
 #' @param error \[`logical(1)`\]\cr
-#' If `check` is not `TRUE`, throw an error?
+#' If `check` is not `TRUE` (or no element in `check` is `TRUE`, if `check` is
+#' a `list`), throw an error?
 #'
 #' @param prefix \[`character(1)`\]\cr
-#' A prefix for the thrown error message, if `check` is not `TRUE` and
-#' `error` is `TRUE`.
+#' A prefix for the thrown error message, only relevant if `error` is `TRUE`.
 #'
 #' @return
-#' `TRUE` if `check` is `TRUE`. If `check` is not `TRUE`, depending on `error`:
+#' `TRUE` if `check` is `TRUE` (or any element in `check` is `TRUE`, if `check`
+#' is a `list`) . Else, depending on `error`:
 #'
 #' - If `error` is `TRUE`, throws an error.
 #' - If `error` is `FALSE`, returns `FALSE`.
@@ -41,6 +45,16 @@
 #'   error = TRUE
 #' )
 #'
+#' ### alternative checks
+#' input_check_response(
+#'   check = list(
+#'     checkmate::check_character(x),
+#'     checkmate::check_character(y)
+#'   ),
+#'   var_name = "x",
+#'   error = TRUE
+#' )
+#'
 #' ### standardized check response
 #' \dontrun{
 #' input_check_response(
@@ -48,15 +62,44 @@
 #'   var_name = "y",
 #'   error = TRUE
 #' )
+#'
+#' input_check_response(
+#'   check = list(
+#'     checkmate::check_flag(x),
+#'     checkmate::check_character(y)
+#'   ),
+#'   var_name = "y",
+#'   error = TRUE
+#' )
 #' }
 
 input_check_response <- function(
-    check, var_name = NULL, error = TRUE, prefix = "Input {.var {var_name}} is bad:"
+    check,
+    var_name = NULL,
+    error = TRUE,
+    prefix = "Input {.var {var_name}} is bad:"
   ) {
-  if (!isTRUE(check)) {
+  if (is.list(check)) {
+    check_true <- sapply(check, isTRUE)
+    check_result <- any(check_true)
+    check_msg <- check[!check_true]
+  } else {
+    check_result <- check_msg <- check
+  }
+  if (!isTRUE(check_result)) {
     if (isTRUE(error)) {
       checkmate::assert_string(prefix, null.ok = TRUE)
-      cli::cli_abort(paste(prefix, "{check}"), call = NULL)
+      if (length(check_msg) > 1) {
+        cli::cli_abort(
+          c(paste(prefix, "Either:"), paste0("- ", check_msg)),
+          call = NULL
+        )
+      } else {
+        cli::cli_abort(
+          paste(prefix, "{check_msg}"),
+          call = NULL
+        )
+      }
     } else {
       FALSE
     }